opsevo-python/pyproject.toml
[project]
name = "opsevo"
version = "0.1.0"
description = "Opsevo-X unified Python backend — device-agnostic AIOps platform"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "psycopg[pool]>=3.2.0",
    "pydantic-settings>=2.5.0",
    "dependency-injector>=4.42.0",
    "structlog>=24.4.0",
    "httpx>=0.27.0",
    "asyncssh>=2.17.0",
    "pysnmp>=7.0.0",
    "apscheduler>=3.10.0",
    "pyjwt>=2.9.0",
    "bcrypt>=4.2.0",
    "python-multipart>=0.0.12",
    "watchfiles>=0.24.0",
    "cryptography>=43.0.0",
    "pyyaml>=6.0.0",
    "typer>=0.12.0",
]

[project.optional-dependencies]
local-embedding = [
    "sentence-transformers>=3.0.0",
]
dev = [
    "hypothesis>=6.112.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[project.scripts]
opsevo-skill = "opsevo.cli.opsevo_skill:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/opsevo"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.hypothesis]
database_backend = "directory"

[tool.ruff]
target-version = "py312"
